Modified the initialization of the ResultRelInfo
One of the problems that we were facing with the SET clause was related to the ResultRelInfo pointing at the wrong RTE via its ri_RangeTableIndex.
The create_entity_result_rel_info() function does not have the capability of setting the ri_RootResultRelInfo to the correct ResultRelInfo node because it does not have access to the ModifyTableState node.
The solution for this was to set the third argument in InitResultRelInfo() to be zero instead of list_length(estate->es_range_table).

Pointed out by Amit Langote